refactor(specfont): rename fontMetrics.js as specFont.js
use this file to set all the font parameters to be used in CSS font properties
