+++ src/utils/specFont.js
@@ -1,16 +1,16 @@
 export const poppins = {
   light: {
+    fontFamily: "'Poppins', Verdana, sans-serif",
+    fontWeight: 300,
     unitsPerEm: 1000,
     xHeight: 546,
     capHeight: 700,
-    fontFamily: "'Poppins', Verdana, sans-serif",
-    fontWeight: 300,
   },
   thin: {
+    fontFamily: "'Poppins', Verdana, sans-serif",
+    fontWeight: 100,
     unitsPerEm: 1000,
     xHeight: 540,
     capHeight: 702,
-    fontFamily: "'Poppins', Verdana, sans-serif",
-    fontWeight: 100,
   },
 }; // Measured with https://opentype.js.org/font-inspector.html
